video_producer_agent/requirements.txt
requests>=2.28.0
pydantic>=1.10.0
python-dotenv>=1.0.0
